updated negative ood warning to display all negative values as a table
